OBE-175 - Session logic slightly wrong
If the user gets login wrong then they are redirected to sessions/new to try
again. Then if they got it right the logic was returning them to the page they
came from, which is now sessions/new and therefore wrong.

As the presence of sessions/new as the referral page is already being detected
it is sensible in this case to redirect back to the root of the site. This is
now what happens.